requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
docutils==0.20.1
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
lockfile==0.12.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
mysql-connector-python==8.3.0
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pid==3.0.4
pillow==10.3.0
protobuf==4.25.3
Pygments==2.17.2
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
rich==13.7.1
service==0.6.0
setproctitle==1.3.3
setuptools==69.2.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
